[Bugfix] HPU logprobs: keep gathered token ids int64 (first-run int32 narrowing corrupted them)
The first logprobs request of a boot returned top_logprobs rows whose token
ids were every other correct id followed by zeros while the logprob values
were right, both on 2026-09-11 and again today after the host-ownership fix,
so it is not a tensor-lifetime race: the ids come out of the device wrong on
the first execution of the gather recipe. The one HPU-specific step in that
recipe is the upstream int64->int32 narrowing of the concatenated indices.
The ids are consumed on the host as Python ints, so drop the narrowing in
both HPU gather paths (patched Sampler.gather_logprobs and the prompt
logprobs gather). Verification is the first-token probe on the next boot;
tools/diag_logprob_probe.py gained a top_logprobs sweep to tell a
first-in-process defect from a first-per-recipe one.

+++ vllm_gaudi/patches.py
@@ -277,7 +277,12 @@ def _hpu_gather_logprobs(
     token_ranks = _sampler_mod.batched_count_greater_than(logprobs, token_logprobs)
     indices = torch.cat((token_ids, topk_indices), dim=1)
     logprobs = torch.cat((token_logprobs, topk_logprobs), dim=1)
-    indices = indices.to(torch.int32)
+    # Upstream narrows ``indices`` to int32 here to save bytes. On HPU the very
+    # first execution of that cat->int32 recipe in a process returned token ids
+    # read at the wrong stride (every other id, then zeros) while the logprob
+    # values were right, producing top_logprobs rows with duplicate tokens and
+    # -inf for the chosen token. The ids are consumed on the host as Python
+    # ints, so keep them int64 and skip the narrowing cast.
     return LogprobsTensors(indices, logprobs, token_ranks)
